+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.9.20+1
+
+* Migrates lifecycle methods (`start`, `stop`, `close`) to Swift.
+* Migrates exposure and focus related methods to Swift.
+* Migrates `receivedImageStreamData` and `reportInitializationState` methods to Swift.
+
 ## 0.9.20
 
 * Fixes incorrect types in image stream events.